phase2: extract prompt_cache.rs and memory_prefetch.rs from bridge_inprocess
- Create turn/prompt_cache.rs (489 lines): PromptCacheConfig, build_system_message,
  annotate_tool_schemas_for_caching, add_message_cache_breakpoint, section_cache_key
  with 18 unit tests
- Create turn/memory_prefetch.rs (307 lines): MemoryPrefetchResult, prefetch_memories,
  merge_memory_results, build_memory_section, extract_entity_tokens, fetch_memories
  with 20 unit tests
- Update server_loop_host.rs imports: bridge_inprocess -> prompt_cache
- Replace inline definitions in bridge_inprocess.rs with pub use re-exports
- Remove duplicate tests from bridge_inprocess.rs test module
- bridge_inprocess.rs: 5920 -> 5247 lines (-673 lines)
- make format + make check + make test-offline: all pass, zero warnings
