+++ rust/crates/runtime/src/turn/bridge_inprocess.rs
@@ -707,285 +707,11 @@ fn rate_limit_cooldown() -> &'static PerModelCooldown {
     COOLDOWN.get_or_init(PerModelCooldown::new)
 }
 
-// ── Cache Configuration (session-latched) ────────────────────────────────────
-// Latched at session start to prevent mid-session flips from busting the cache.
-
-/// Prompt cache configuration, latched once per session.
-///
-/// Instead of reading `MO_PROMPT_CACHE_DISABLED` on every LLM call (which risks
-/// mid-session env var changes busting the KV cache), this struct captures the
-/// config at session init and is passed to all cache-related functions.
-#[derive(Debug, Clone)]
-pub struct PromptCacheConfig {
-    /// Whether cache_control annotations are enabled for Anthropic.
-    pub cache_enabled: bool,
-    /// Whether the provider supports cache_control (Anthropic/Claude).
-    pub is_anthropic: bool,
-}
-
-impl PromptCacheConfig {
-    /// Latch config from environment and provider info. Call once at session start.
-    pub fn latch(provider: &str, model_name: &str) -> Self {
-        let cache_enabled = !std::env::var("MO_PROMPT_CACHE_DISABLED")
-            .is_ok_and(|v| v == "1" || v.eq_ignore_ascii_case("true"));
-        let is_anthropic = provider == "anthropic" || model_name.contains("claude");
-        Self {
-            cache_enabled,
-            is_anthropic,
-        }
-    }
-
-    /// Convenience: should we emit cache_control annotations?
-    pub fn should_annotate(&self) -> bool {
-        self.cache_enabled && self.is_anthropic
-    }
-}
-
-impl Default for PromptCacheConfig {
-    fn default() -> Self {
-        Self {
-            cache_enabled: true,
-            is_anthropic: false,
-        }
-    }
-}
-
-// ── System Prompt Cache ──────────────────────────────────────────────────────
-// Two-level cache for static/dynamic prompt boundary:
-// - Global+Session sections are cached by (tool_names, task_type, confidence) — stable within a session
-// - Per-turn profile_desc is NOT cached (changes every turn with skills/memory/environment)
-use std::sync::Mutex;
-
-/// Cached prompt sections (Global + Session scoped).
-struct CachedSections {
-    /// Concatenated text of Global+Session sections (for non-Anthropic providers).
-    text: String,
-    /// Individual sections with scope metadata (for Anthropic cache_control).
-    sections: Vec<prompts::PromptSection>,
-}
-
-fn section_cache() -> &'static Mutex<HashMap<u64, CachedSections>> {
-    static CACHE: OnceLock<Mutex<HashMap<u64, CachedSections>>> = OnceLock::new();
-    CACHE.get_or_init(|| Mutex::new(HashMap::new()))
-}
-
-fn section_cache_key(tool_names: &[&str], task_type: Option<&str>, confidence: f64) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    for name in tool_names {
-        name.hash(&mut hasher);
-    }
-    task_type.unwrap_or("none").hash(&mut hasher);
-    let bucket = if confidence < 0.3 { "low" } else { "normal" };
-    bucket.hash(&mut hasher);
-    hasher.finish()
-}
-
-/// Build the system message Value for the LLM API.
-///
-/// For Anthropic providers: uses content array with `cache_control` on stable sections,
-/// enabling server-side KV cache reuse across turns.
-///
-/// For other providers (OpenAI, DeepSeek, etc.): uses a single content string.
-/// Build the system message(s) for the LLM API.
-///
-/// Returns `(primary, dynamic)`:
-/// - **Anthropic**: `primary` is a multi-block content array with `cache_control` on stable
-///   sections and dynamic profile appended without cache markers. `dynamic` is `None`.
-/// - **OpenAI / other**: `primary` contains only the **stable** text (cacheable prefix).
-///   `dynamic` holds a second system message with the per-turn profile/hints, or `None`
-///   if there is nothing dynamic. This split enables OpenAI's automatic prefix caching:
-///   the stable message stays identical across turns so the provider can reuse the KV cache.
-pub(crate) fn build_system_message(
-    tool_names: &[&str],
-    profile_desc: &str,
-    confidence: f64,
-    task_type: Option<&str>,
-    cache_cfg: &PromptCacheConfig,
-) -> (Value, Option<Value>, Vec<prompts::PromptSection>) {
-    let key = section_cache_key(tool_names, task_type, confidence);
-
-    // Try cache for the stable (Global + Session) sections
-    let cached = if let Ok(cache) = section_cache().lock() {
-        cache
-            .get(&key)
-            .map(|c| (c.text.clone(), c.sections.clone()))
-    } else {
-        None
-    };
-
-    let (stable_text, sections) = cached.unwrap_or_else(|| {
-        // Build all sections (profile_desc is "" for cache — we'll append it separately)
-        let all = prompts::build_system_prompt_sections(tool_names, "", confidence, task_type);
-        // Only cache Global + Session sections (not None-scoped profile)
-        let stable: Vec<prompts::PromptSection> = all
-            .into_iter()
-            .filter(|s| s.scope != prompts::CacheScope::None)
-            .collect();
-        let text = stable
-            .iter()
-            .map(|s| s.text.as_str())
-            .collect::<Vec<_>>()
-            .join("");
-
-        if let Ok(mut cache) = section_cache().lock() {
-            if cache.len() > 32 {
-                cache.clear();
-            }
-            cache.insert(
-                key,
-                CachedSections {
-                    text: text.clone(),
-                    sections: stable.clone(),
-                },
-            );
-        }
-        (text, stable)
-    });
-
-    let is_anthropic = cache_cfg.is_anthropic;
-
-    // Build complete sections list (stable + dynamic) for trace.
-    let append_dynamic = |mut secs: Vec<prompts::PromptSection>| -> Vec<prompts::PromptSection> {
-        if !profile_desc.is_empty() {
-            secs.push(prompts::PromptSection {
-                text: profile_desc.to_string(),
-                scope: prompts::CacheScope::None,
-            });
-        }
-        secs
-    };
-
-    if is_anthropic {
-        // Anthropic: multi-block content with cache_control on stable sections.
-        // Even via OpenAI-compatible proxies, many forward cache_control to the
-        // native Messages API. Proxies that don't will simply ignore the field.
-        //
-        // Cache strategy:
-        //   Place cache_control on the LAST block of each scope group.
-        //   Anthropic allows up to 4 breakpoints per request — we use at most 2
-        //   (last Global, last Session). The provider caches the prefix up to
-        //   each breakpoint.
-        //
-        //   Global  → scope:"global" + ttl:"1h"  (shared across all sessions/orgs)
-        //   Session → ttl:"1h"                    (stable within a session)
-        //   None    → no cache_control             (changes every turn)
-        let cache_disabled = !cache_cfg.cache_enabled;
-
-        // Find the last index of each scope group for breakpoint placement
-        let last_global = sections
-            .iter()
-            .rposition(|s| s.scope == prompts::CacheScope::Global);
-        let last_session = sections
-            .iter()
-            .rposition(|s| s.scope == prompts::CacheScope::Session);
-
-        let mut blocks: Vec<Value> = Vec::with_capacity(sections.len() + 1);
-        for (i, section) in sections.iter().enumerate() {
-            let cc = if cache_disabled {
-                None
-            } else if Some(i) == last_global {
-                Some(json!({"type": "ephemeral", "scope": "global", "ttl": "1h"}))
-            } else if Some(i) == last_session {
-                Some(json!({"type": "ephemeral", "ttl": "1h"}))
-            } else {
-                None
-            };
-            let mut block = json!({
-                "type": "text",
-                "text": section.text,
-            });
-            if let Some(cc) = cc {
-                block["cache_control"] = cc;
-            }
-            blocks.push(block);
-        }
-        // Dynamic section (profile + per-turn hints) — no cache_control
-        if !profile_desc.is_empty() {
-            blocks.push(json!({
-                "type": "text",
-                "text": profile_desc,
-            }));
-        }
-        // Anthropic: everything in one message (cache_control breakpoints handle caching)
-        (
-            json!({
-                "role": "system",
-                "content": blocks,
-            }),
-            None,
-            append_dynamic(sections),
-        )
-    } else {
-        // OpenAI-compatible: split stable / dynamic into separate system messages
-        // so the stable prefix is identical across turns and the provider can reuse
-        // its automatic KV cache.
-        let primary = json!({
-            "role": "system",
-            "content": stable_text,
-        });
-        let dynamic = if profile_desc.is_empty() {
-            None
-        } else {
-            Some(json!({
-                "role": "system",
-                "content": profile_desc,
-            }))
-        };
-        (primary, dynamic, append_dynamic(sections))
-    }
-}
-
-/// Add `cache_control` to the last tool schema for Anthropic caching.
-///
-/// Anthropic allows only 4 cache_control breakpoints per request. Our allocation:
-/// - System prompt: 2 breakpoints (global scope + session scope)
-/// - Tools: 1 breakpoint (last tool only)
-/// - Messages: 1 breakpoint (last message)
-///
-/// Prioritizes message history caching for multi-turn conversations
-/// while still caching the stable tool prefix.
-pub(crate) fn annotate_tool_schemas_for_caching(
-    tools: &mut [Value],
-    cache_cfg: &PromptCacheConfig,
-) {
-    if !cache_cfg.should_annotate() || tools.is_empty() {
-        return;
-    }
-    // Mark only the last tool — this creates a single cache covering all tools.
-    // Since tools rarely change mid-conversation, this is usually a cache hit.
-    let last_idx = tools.len() - 1;
-    tools[last_idx]["cache_control"] = json!({"type": "ephemeral", "ttl": "1h"});
-}
-
-/// Add a cache breakpoint on the last conversation message for Anthropic.
-/// This enables turn-to-turn KV cache reuse for the conversation prefix.
-pub(crate) fn add_message_cache_breakpoint(messages: &mut [Value], cache_cfg: &PromptCacheConfig) {
-    if !cache_cfg.should_annotate() || messages.is_empty() {
-        return;
-    }
-    // Find the last non-system message and add cache_control to it
-    if let Some(last) = messages.iter_mut().rev().find(|m| {
-        m.get("role")
-            .and_then(Value::as_str)
-            .is_some_and(|r| r != "system")
-    }) {
-        // If content is a string, convert to array format for cache_control
-        if last.get("content").is_some_and(Value::is_string) {
-            let text = last["content"].as_str().unwrap_or_default().to_string();
-            last["content"] = json!([{
-                "type": "text",
-                "text": text,
-                "cache_control": {"type": "ephemeral"},
-            }]);
-        } else if let Some(arr) = last.get_mut("content").and_then(Value::as_array_mut) {
-            // Content is already an array — add cache_control to last element
-            if let Some(last_block) = arr.last_mut() {
-                last_block["cache_control"] = json!({"type": "ephemeral"});
-            }
-        }
-    }
-}
+// ── Prompt caching — delegated to turn::prompt_cache ─────────────────────────
+pub use super::prompt_cache::PromptCacheConfig;
+pub(crate) use super::prompt_cache::{
+    add_message_cache_breakpoint, annotate_tool_schemas_for_caching, build_system_message,
+};
 
 fn bridge_llm_cancel(cc: &Option<Arc<CancellationToken>>) -> LlmCancel<'_> {
     match cc.as_ref() {
@@ -3037,150 +2763,8 @@ fn classify_llm_error(msg: &str) -> astra_core::ErrorKind {
     crate::turn::llm_client::classify_llm_error(msg)
 }
 
-/// Result of a memory prefetch operation.
-#[derive(Debug, Default)]
-pub struct MemoryPrefetchResult {
-    pub section: Option<String>,
-    pub items: usize,
-    pub preview: Vec<String>,
-    pub fetch_ms: i64,
-}
-
-/// Prefetch memories relevant to the user message via hybrid retrieval.
-/// Sends two queries (full message + entity tokens), merges and deduplicates.
-pub async fn prefetch_memories(
-    mem_url: &str,
-    mem_key: &str,
-    user_msg: &str,
-    user_id: &str,
-    top_k: u32,
-) -> MemoryPrefetchResult {
-    if mem_key.is_empty() || user_msg.trim().is_empty() {
-        return MemoryPrefetchResult::default();
-    }
-    let started = Instant::now();
-    let entity_query = extract_entity_tokens(user_msg);
-    let trimmed_msg = user_msg.trim();
-
-    // Parallel fetch: full message retrieval + entity-keyword retrieval via tokio::join!
-    // Saves one round-trip latency (~50-200ms) compared to sequential.
-    let do_entity = !entity_query.is_empty() && entity_query != trimmed_msg;
-    let (full_result, entity_result) = tokio::join!(
-        fetch_memories(mem_url, mem_key, trimmed_msg, user_id, top_k),
-        async {
-            if do_entity {
-                fetch_memories(mem_url, mem_key, &entity_query, user_id, top_k).await
-            } else {
-                String::new()
-            }
-        }
-    );
-    let merged = merge_memory_results(&[&full_result, &entity_result]);
-    let fetch_ms = started.elapsed().as_millis() as i64;
-    let preview = merged.iter().take(3).map(|l| l.to_string()).collect();
-    let items = merged.len();
-    let section = build_memory_section(&merged);
-    MemoryPrefetchResult {
-        section,
-        items,
-        preview,
-        fetch_ms,
-    }
-}
-
-/// Merge and deduplicate memory results from multiple retrieval queries.
-fn merge_memory_results(results: &[&str]) -> Vec<String> {
-    let mut seen = std::collections::HashSet::new();
-    let mut merged = Vec::new();
-    for result in results {
-        for line in result.lines() {
-            let trimmed = line.trim();
-            if !trimmed.is_empty() && seen.insert(trimmed.to_string()) {
-                merged.push(trimmed.to_string());
-            }
-        }
-    }
-    merged
-}
-
-/// Build the memory section for the profile block.
-/// Returns None if no memories matched.
-fn build_memory_section(merged_lines: &[String]) -> Option<String> {
-    if merged_lines.is_empty() {
-        return None;
-    }
-    let refs: Vec<&str> = merged_lines.iter().map(|s| s.as_str()).collect();
-    let formatted = crate::prompts::memory_proto::format_for_llm(&refs);
-    if !formatted.is_empty() {
-        Some(format!("## User Memories\n{formatted}"))
-    } else {
-        Some(format!("## User Memories\n{}", merged_lines.join("\n")))
-    }
-}
-
-/// Extract non-CJK, non-punctuation tokens from a message for keyword-based retrieval.
-/// General purpose: works for any mixed-language input, not specific to any domain.
-fn extract_entity_tokens(msg: &str) -> String {
-    let mut tokens = Vec::new();
-    let mut current = String::new();
-    for ch in msg.chars() {
-        if ch.is_ascii_alphanumeric() || ch == '-' || ch == '_' {
-            current.push(ch);
-        } else {
-            if current.len() >= 3 {
-                tokens.push(current.clone());
-            }
-            current.clear();
-        }
-    }
-    if current.len() >= 3 {
-        tokens.push(current);
-    }
-    tokens.join(" ")
-}
-
-/// Fetch memories from Memoria HTTP API. Returns joined content string.
-async fn fetch_memories(
-    base_url: &str,
-    api_key: &str,
-    query: &str,
-    user_id: &str,
-    top_k: u32,
-) -> String {
-    let client = reqwest::Client::builder()
-        .no_proxy()
-        .build()
-        .unwrap_or_else(|_| reqwest::Client::new());
-    let mut payload = serde_json::json!({"query": query, "top_k": top_k});
-    if !user_id.is_empty() {
-        payload["session_id"] = serde_json::Value::String(user_id.to_string());
-        payload["user_id"] = serde_json::Value::String(user_id.to_string());
-    }
-    let resp = match client
-        .post(format!("{base_url}/v1/memories/retrieve"))
-        .header("Authorization", format!("Bearer {api_key}"))
-        .json(&payload)
-        .send()
-        .await
-    {
-        Ok(r) => r,
-        Err(e) => {
-            astra_core::agent_error!("memory", "fetch error: {e:#}");
-            return String::new();
-        }
-    };
-    if !resp.status().is_success() {
-        return String::new();
-    }
-    let arr = match resp.json::<Vec<serde_json::Value>>().await {
-        Ok(a) => a,
-        Err(_) => return String::new(),
-    };
-    arr.iter()
-        .filter_map(|m| m.get("content").and_then(|c| c.as_str()))
-        .collect::<Vec<_>>()
-        .join("\n")
-}
+// ── Memory prefetch — delegated to turn::memory_prefetch ─────────────────────
+pub use super::memory_prefetch::{MemoryPrefetchResult, prefetch_memories};
 
 /// Test-accessible wrapper around private schema pruning — used by integration
 /// tests that need to verify progressive schema detail levels.
@@ -3197,117 +2781,6 @@ pub mod bridge_inprocess_test_helpers {
 mod tests {
     use super::*;
     use crate::turn::turn_guard::TurnGuard;
-
-    #[test]
-    fn extract_entity_tokens_from_mixed_language() {
-        assert_eq!(extract_entity_tokens("memoria 最新的ci?"), "memoria");
-        assert_eq!(
-            extract_entity_tokens("matrixone latest pr"),
-            "matrixone latest"
-        );
-        assert_eq!(extract_entity_tokens("你好"), "");
-        assert_eq!(
-            extract_entity_tokens("check astra status"),
-            "check astra status"
-        );
-    }
-
-    // ── merge_memory_results ──────────────────────────────────────────────────
-
-    #[test]
-    fn merge_deduplicates_across_queries() {
-        let r1 = "[@fact/semantic] memoria is matrixorigin/memoria\nsome other fact";
-        let r2 = "[@fact/semantic] memoria is matrixorigin/memoria\nnew fact";
-        let merged = merge_memory_results(&[r1, r2]);
-        assert_eq!(
-            merged.len(),
-            3,
-            "duplicate should be removed, got: {merged:?}"
-        );
-        assert!(merged.contains(&"[@fact/semantic] memoria is matrixorigin/memoria".to_string()));
-        assert!(merged.contains(&"some other fact".to_string()));
-        assert!(merged.contains(&"new fact".to_string()));
-    }
-
-    #[test]
-    fn merge_skips_empty_lines() {
-        let r1 = "line1\n\n\nline2";
-        let r2 = "";
-        let merged = merge_memory_results(&[r1, r2]);
-        assert_eq!(merged, vec!["line1", "line2"]);
-    }
-
-    #[test]
-    fn merge_empty_inputs() {
-        assert!(merge_memory_results(&["", ""]).is_empty());
-        assert!(merge_memory_results(&[]).is_empty());
-    }
-
-    // ── build_memory_section ──────────────────────────────────────────────────
-
-    #[test]
-    fn build_memory_section_returns_none_for_empty() {
-        assert!(build_memory_section(&[]).is_none());
-    }
-
-    #[test]
-    fn build_memory_section_includes_header() {
-        let lines = vec!["[@pref/active] memoria = matrixorigin/Memoria".to_string()];
-        let section = build_memory_section(&lines).unwrap();
-        assert!(section.starts_with("## User Memories"), "got: {section}");
-    }
-
-    #[test]
-    fn section_cache_key_varies_by_tools_and_task() {
-        let key1 = section_cache_key(&["bash"], Some("implementation"), 0.8);
-        let key2 = section_cache_key(&["bash", "read_file"], Some("implementation"), 0.8);
-        let key3 = section_cache_key(&["bash"], Some("debugging"), 0.8);
-        let key4 = section_cache_key(&["bash"], Some("implementation"), 0.2);
-        assert_ne!(key1, key2, "different tools should differ");
-        assert_ne!(key1, key3, "different task types should differ");
-        assert_ne!(key1, key4, "different confidence buckets should differ");
-    }
-
-    #[test]
-    fn build_memory_section_formats_structured_entries() {
-        let lines = vec!["[@pref/active] dark mode preferred".to_string()];
-        let section = build_memory_section(&lines).unwrap();
-        assert!(
-            section.contains("Preferences"),
-            "structured entries should be grouped, got: {section}"
-        );
-    }
-
-    #[test]
-    fn build_memory_section_handles_unstructured() {
-        let lines = vec!["just a plain memory without tags".to_string()];
-        let section = build_memory_section(&lines).unwrap();
-        assert!(section.contains("just a plain memory"), "got: {section}");
-    }
-
-    // ── entity + merge integration ────────────────────────────────────────────
-
-    #[test]
-    fn entity_query_differs_from_mixed_language_input() {
-        let msg = "memoria 最新的ci?";
-        let entity = extract_entity_tokens(msg);
-        assert_ne!(
-            entity,
-            msg.trim(),
-            "entity query should differ for mixed-language"
-        );
-        assert_eq!(entity, "memoria");
-    }
-
-    #[test]
-    fn entity_query_same_for_pure_ascii() {
-        let msg = "memoria latest ci";
-        let entity = extract_entity_tokens(msg);
-        assert_eq!(
-            entity, "memoria latest",
-            "pure ASCII: entity ≈ original (minus short words)"
-        );
-    }
 
     #[test]
     fn count_inprocess_persisted_events_skips_failed_tool_events() {
@@ -4131,68 +3604,6 @@ mod tests {
         }
     }
 
-    // ── Section cache eviction test ────────────────────────────────────
-
-    #[test]
-    fn section_cache_evicts_after_capacity() {
-        let _lock = CACHE_ENV_MUTEX.lock().unwrap();
-
-        // Clear any pre-existing cache entries
-        if let Ok(mut cache) = section_cache().lock() {
-            cache.clear();
-        }
-
-        // Fill cache to 33 entries (> 32 threshold), then one more triggers clear
-        for i in 0..34 {
-            let tool_name = format!("tool_{i}");
-            let tools: Vec<&str> = vec![tool_name.as_str()];
-            let (_msg, _, _) = build_system_message(
-                &tools,
-                "",
-                0.8,
-                None,
-                &PromptCacheConfig::latch("openai", "gpt-4"),
-            );
-        }
-
-        let cache_size = section_cache().lock().unwrap().len();
-        // After 34th insert: cache had 33 entries (> 32), cleared, then 34th re-added → 1
-        assert!(
-            cache_size <= 2,
-            "cache should have been evicted: size={cache_size}, expected ≤2"
-        );
-
-        // Clean up
-        section_cache().lock().unwrap().clear();
-    }
-
-    #[test]
-    fn section_cache_key_deterministic_for_same_inputs() {
-        let k1 = section_cache_key(&["bash", "read_file"], Some("debug"), 0.5);
-        let k2 = section_cache_key(&["bash", "read_file"], Some("debug"), 0.5);
-        assert_eq!(k1, k2, "same inputs should produce same key");
-    }
-
-    #[test]
-    fn section_cache_key_differs_for_different_tools() {
-        let k1 = section_cache_key(&["bash"], None, 0.8);
-        let k2 = section_cache_key(&["read_file"], None, 0.8);
-        assert_ne!(k1, k2, "different tools should produce different keys");
-    }
-
-    #[test]
-    fn section_cache_key_low_confidence_bucketed() {
-        // confidence < 0.3 → "low" bucket, >= 0.3 → "normal" bucket
-        let low = section_cache_key(&["bash"], None, 0.1);
-        let normal = section_cache_key(&["bash"], None, 0.5);
-        assert_ne!(low, normal, "low vs normal confidence should differ");
-
-        // Both in normal bucket → same key
-        let n1 = section_cache_key(&["bash"], None, 0.5);
-        let n2 = section_cache_key(&["bash"], None, 0.9);
-        assert_eq!(n1, n2, "both normal confidence should be same bucket");
-    }
-
     // ── Message breakpoint edge cases ──────────────────────────────────
 
     #[test]
@@ -4898,90 +4309,6 @@ mod tests {
         assert_eq!(event["type"], "session_info");
         assert_eq!(event["session_id"], "sess-1");
         assert_eq!(event["run_id"], "run-1");
-    }
-
-    #[test]
-    fn extract_entity_tokens_empty_string() {
-        assert_eq!(extract_entity_tokens(""), "");
-    }
-
-    #[test]
-    fn extract_entity_tokens_short_words_filtered() {
-        // Words < 3 chars are dropped
-        assert_eq!(extract_entity_tokens("a b cd ef"), "");
-    }
-
-    #[test]
-    fn extract_entity_tokens_preserves_long_tokens() {
-        assert_eq!(extract_entity_tokens("hello world"), "hello world");
-    }
-
-    #[test]
-    fn extract_entity_tokens_special_chars_split() {
-        assert_eq!(
-            extract_entity_tokens("user.name@domain.com"),
-            "user name domain com"
-        );
-    }
-
-    #[test]
-    fn extract_entity_tokens_hyphens_and_underscores_kept() {
-        assert_eq!(extract_entity_tokens("my-app_v2"), "my-app_v2");
-    }
-
-    #[test]
-    fn extract_entity_tokens_unicode_chars_as_delimiters() {
-        // CJK chars and emoji act as delimiters
-        let result = extract_entity_tokens("hello你好world");
-        // 'hello' is 5 chars, '你好' splits, 'world' is 5 chars
-        assert_eq!(result, "hello world");
-    }
-
-    #[test]
-    fn extract_entity_tokens_only_special_chars() {
-        assert_eq!(extract_entity_tokens("!@#$%^&*()"), "");
-    }
-
-    #[test]
-    fn prefetch_memories_empty_key_returns_default() {
-        let rt = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap();
-        let result = rt.block_on(prefetch_memories(
-            "http://localhost",
-            "",
-            "query",
-            "user1",
-            5,
-        ));
-        assert_eq!(result.items, 0);
-        assert!(result.section.is_none());
-    }
-
-    #[test]
-    fn prefetch_memories_whitespace_message_returns_default() {
-        let rt = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap();
-        let result = rt.block_on(prefetch_memories(
-            "http://localhost",
-            "key",
-            "   ",
-            "user1",
-            5,
-        ));
-        assert_eq!(result.items, 0);
-    }
-
-    #[test]
-    fn memory_prefetch_result_default() {
-        let r = MemoryPrefetchResult::default();
-        assert!(r.section.is_none());
-        assert_eq!(r.items, 0);
-        assert!(r.preview.is_empty());
-        assert_eq!(r.fetch_ms, 0);
     }
 
     // ── render_sse tests ────────────────────────────────────────────────
