+++ rust/crates/runtime/src/turn/mod.rs
@@ -82,6 +82,7 @@ pub mod interruption;
 pub(crate) mod llm_client;
 pub mod llm_request_dump;
 pub mod loop_dispatcher;
+pub mod memory_prefetch;
 pub mod microcompact;
 pub mod observer;
 pub mod parallel_tool_exec;
@@ -90,6 +91,7 @@ pub mod persist;
 pub mod persist_inputs;
 pub mod prepare_turn_explain_text;
 pub mod progress_display;
+pub mod prompt_cache;
 pub mod quality;
 pub mod refresh;
 pub mod response_guard;
